+++ crates/bcr-ebill-persistence/src/db/company_chain.rs
@@ -5,8 +5,9 @@ use super::{
 use crate::{
     company::CompanyChainStoreApi,
     constants::{
-        DB_BLOCK_ID, DB_COMPANY_ID, DB_DATA, DB_HASH, DB_OP_CODE, DB_PREVIOUS_HASH, DB_PUBLIC_KEY,
-        DB_SIGNATORY_NODE_ID, DB_SIGNATURE, DB_TABLE, DB_TIMESTAMP,
+        DB_BLOCK_ID, DB_COMPANY_ID, DB_DATA, DB_HASH, DB_OP_CODE, DB_PLAINTEXT_HASH,
+        DB_PREVIOUS_HASH, DB_PUBLIC_KEY, DB_SIGNATORY_NODE_ID, DB_SIGNATURE, DB_TABLE,
+        DB_TIMESTAMP,
     },
 };
 use async_trait::async_trait;
@@ -22,6 +23,7 @@ use serde::{Deserialize, Serialize};
 const CREATE_BLOCK_QUERY: &str = r#"CREATE type::table($table) CONTENT {
                                     company_id: $company_id,
                                     block_id: $block_id,
+                                    plaintext_hash: $plaintext_hash,
                                     hash: $hash,
                                     previous_hash: $previous_hash,
                                     signature: $signature,
@@ -49,6 +51,7 @@ impl SurrealCompanyChainStore {
         bindings.add(DB_TABLE, Self::TABLE)?;
         bindings.add(DB_COMPANY_ID, entity.company_id)?;
         bindings.add(DB_BLOCK_ID, entity.block_id)?;
+        bindings.add(DB_PLAINTEXT_HASH, entity.plaintext_hash)?;
         bindings.add(DB_HASH, entity.hash)?;
         bindings.add(DB_PREVIOUS_HASH, entity.previous_hash)?;
         bindings.add(DB_SIGNATURE, entity.signature)?;
@@ -187,6 +190,7 @@ impl CompanyChainStoreApi for SurrealCompanyChainStore {
 pub struct CompanyBlockDb {
     pub company_id: NodeId,
     pub block_id: u64,
+    pub plaintext_hash: String,
     pub hash: String,
     pub previous_hash: String,
     pub signature: String,
@@ -202,6 +206,7 @@ impl From<CompanyBlockDb> for CompanyBlock {
         Self {
             company_id: value.company_id,
             id: value.block_id,
+            plaintext_hash: value.plaintext_hash,
             hash: value.hash,
             timestamp: value.timestamp,
             data: value.data,
@@ -219,6 +224,7 @@ impl From<&CompanyBlock> for CompanyBlockDb {
         Self {
             company_id: value.company_id.clone(),
             block_id: value.id,
+            plaintext_hash: value.plaintext_hash.clone(),
             hash: value.hash.clone(),
             previous_hash: value.previous_hash.clone(),
             signature: value.signature.clone(),