+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.18
+
+* Add `plaintext_hash` to Identity, Company and Bill Blocks, which is a hash over the plaintext data
+    * (breaks all chains in the DB)
+
 # 0.3.17
 
 * Changed minted proofs token format from cashu Token v3 to BitcrB (v4)